requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async==0.6.2
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.22.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.9.0
ujson==5.9.0
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
gunicorn

